backend/requirements.txt
fastapi[all]==0.95.1
pymongo==4.3.3
pydantic==1.10.7
pydantic[email]==1.10.7
uvicorn==0.22.0
python-dotenv==0.20.0
PyPDF2==3.0.1
langchain==0.2.6
langchain_community==0.2.6
openai==0.28